Fix chrome binary path for Windows OS
The binary path under Windows are usually like the following one.
C:\Program Files (x86)\Google\Chrome\Application\chrome.exe

The problem is that there is a space in the path and it causes
the glued shell command to be split accidentally. So I guess it
would be good to always use a quoted path if possible.

Signed-off-by: Jack Cherng <jfcherng@gmail.com>